gst-vaapi: move common code to lib
